requirements.txt
Pillow
Requests
streamlit
torch
torchvision
